Rename libManager to fileLoader
